Add a new type of PerfEvent
This adds VendorEvent.
The goal of that event is give an easy and generic way to
add a PerfEvent that will have not standard properties
(implementation defined).
This will left possiblity to see custom events in tools using
the class Perf. The only change in the tools is to request
for a vendor event instead of regular event.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>